requirements.txt
streamlit
pillow
pandas
numpy
google-generativeai
google-cloud-aiplatform